+++ lib/commercetools-api/src/Models/OrderEdit/StagedOrderModel.php
@@ -17,6 +17,8 @@ use Commercetools\Api\Models\Cart\DirectDiscountCollection;
 use Commercetools\Api\Models\Cart\DiscountCodeInfoCollection;
 use Commercetools\Api\Models\Cart\DiscountOnTotalPrice;
 use Commercetools\Api\Models\Cart\DiscountOnTotalPriceModel;
+use Commercetools\Api\Models\Cart\DiscountTypeCombination;
+use Commercetools\Api\Models\Cart\DiscountTypeCombinationModel;
 use Commercetools\Api\Models\Cart\LineItemCollection;
 use Commercetools\Api\Models\Cart\ShippingCollection;
 use Commercetools\Api\Models\Cart\ShippingInfo;
@@ -340,6 +342,12 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
     protected $returnInfo;
 
     /**
+     *
+     * @var ?DiscountTypeCombination
+     */
+    protected $discountTypeCombination;
+
+    /**
      * @deprecated
      * @var ?int
      */
@@ -420,6 +428,7 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
         ?StateReference $state = null,
         ?SyncInfoCollection $syncInfo = null,
         ?ReturnInfoCollection $returnInfo = null,
+        ?DiscountTypeCombination $discountTypeCombination = null,
         ?int $lastMessageSequenceNumber = null,
         ?CustomFields $custom = null,
         ?DateTimeImmutable $completedAt = null,
@@ -472,6 +481,7 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
         $this->state = $state;
         $this->syncInfo = $syncInfo;
         $this->returnInfo = $returnInfo;
+        $this->discountTypeCombination = $discountTypeCombination;
         $this->lastMessageSequenceNumber = $lastMessageSequenceNumber;
         $this->custom = $custom;
         $this->completedAt = $completedAt;
@@ -1448,6 +1458,27 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
     }
 
     /**
+     * <p>Indicates if a combination of discount types can apply on an Order.</p>
+     *
+     *
+     * @return null|DiscountTypeCombination
+     */
+    public function getDiscountTypeCombination()
+    {
+        if (is_null($this->discountTypeCombination)) {
+            /** @psalm-var stdClass|array<string, mixed>|null $data */
+            $data = $this->raw(self::FIELD_DISCOUNT_TYPE_COMBINATION);
+            if (is_null($data)) {
+                return null;
+            }
+            $className = DiscountTypeCombinationModel::resolveDiscriminatorClass($data);
+            $this->discountTypeCombination = $className::of($data);
+        }
+
+        return $this->discountTypeCombination;
+    }
+
+    /**
      * <p>Internal-only field.</p>
      *
      * @deprecated
@@ -1922,6 +1953,14 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
     public function setReturnInfo(?ReturnInfoCollection $returnInfo): void
     {
         $this->returnInfo = $returnInfo;
+    }
+
+    /**
+     * @param ?DiscountTypeCombination $discountTypeCombination
+     */
+    public function setDiscountTypeCombination(?DiscountTypeCombination $discountTypeCombination): void
+    {
+        $this->discountTypeCombination = $discountTypeCombination;
     }
 
     /**